# doc/api-spec.yaml
openapi: 3.0.0
paths:
...
  /artist:
    get:
      operationId: ArtistsController_findAll
      summary: Gets all artists
      parameters: []
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Artist"
        "401":
          description: Access token is missing or invalid
      tags: &a2
        - Artist
    post:
      operationId: ArtistsController_create
      summary: Add new artist
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArtistDto"
      responses:
        "201":
          description: The user has been created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
        "400":
          description: Body does not contain required fields
        "401":
          description: Access token is missing or invalid
      tags: *a2
  "/artist/{artistId}":
    get:
      operationId: ArtistsController_findOne
      summary: Get single artist by ID
      parameters:
        - name: artistId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
        "400":
          description: id is invalid (not uuid)
        "401":
          description: Access token is missing or invalid
        "404":
          description: Not found
      tags: *a2
    put:
      operationId: ArtistsController_update
      summary: Update artist information by ID
      parameters:
        - name: artistId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        description: Have to contain at least one field
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateArtistDto"
            examples:
              full:
                value:
                  name: James Bond
                  grammy: true
              partial:
                value:
                  grammy: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                title: Artist
                properties:
                  id:
                    type: string
                    format: uuid
                    example: 98555ff3-6770-40ed-bf5f-2344713cd3ec
                  name:
                    type: string
                    example: Freddie Mercury
                  grammy:
                    type: boolean
                    example: true
        "400":
          description: id is invalid (not uuid) or Body is empty
        "401":
          description: Access token is missing or invalid
        "404":
          description: Not found
      tags: *a2
    delete:
      operationId: ArtistsController_remove
      summary: Delete artist by ID from library
      parameters:
        - name: artistId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        "204":
          description: The artist has been deleted
        "400":
          description: id is invalid (not uuid)
        "401":
          description: Access token is missing or invalid
        "404":
          description: Not found
      tags: *a2
  /album:
    get:
      operationId: AlbumsController_findAll
      summary: Gets all albums
      parameters: []
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Album"
        "401":
          description: Access token is missing or invalid
      tags: &a3
        - Album
    post:
      operationId: AlbumsController_create
      summary: Add new album
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlbumDto"
      responses:
        "201":
          description: The album has been created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"
        "400":
          description: Body does not contain required fields
        "401":
          description: Access token is missing or invalid
      tags: *a3
  "/album/{albumId}":
    get:
      operationId: AlbumsController_findOne
      summary: Get single album by ID
      parameters:
        - name: albumId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"
        "400":
          description: id is invalid (not uuid)
        "401":
          description: Access token is missing or invalid
        "404":
          description: Not found
      tags: *a3
    put:
      operationId: AlbumsController_update
      summary: Update album information by ID
      parameters:
        - name: albumId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        description: Have to contain at least one field
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAlbumDto"
            examples:
              full:
                value:
                  name: Diamond
                  year: 2010
                  artistId: d3fe6fad-5dba-4ab5-be5e-526c336d6114
              partial:
                value:
                  name: Diamond
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                title: Album
                properties:
                  id:
                    type: string
                    format: uuid
                    example: 98555ff3-6770-40ed-bf5f-2344713cd3ec
                  name:
                    type: string
                    example: Diamond
                  year:
                    type: integer
                    example: 2010
                  artistId:
                    type: string
                    format: uuid
                    example: d3fe6fad-5dba-4ab5-be5e-526c336d6114
        "400":
          description: id is invalid (not uuid) or Body is empty
        "401":
          description: Access token is missing or invalid
        "404":
          description: Not found
      tags: *a3
    delete:
      operationId: AlbumsController_remove
      summary: Delete album by ID from library
      parameters:
        - name: albumId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        "204":
          description: The album has been deleted
        "400":
          description: id is invalid (not uuid)
        "401":
          description: Access token is missing or invalid
        "404":
          description: Not found
      tags: *a3
...
  "/favs/album/{id}":
    post:
      operationId: FavoritesController_addAlbum
      summary: Add album to the favorites
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        "201":
          description: Added successfully
        "400":
          description: id is invalid (not uuid)
        "401":
          description: Access token is missing or invalid
        "422":
          description: Album with id doesn't exist
      tags: *a4
    delete:
      operationId: FavoritesController_deleteAlbum
      summary: Delete album from favorites
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        "204":
          description: Deleted successfully
        "400":
          description: id is invalid (not uuid)
        "401":
          description: Access token is missing or invalid
        "404":
          description: Album was not found
      tags: *a4
  "/favs/artist/{id}":
    post:
      operationId: FavoritesController_addArtist
      summary: Add artist to the favorites
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        "201":
          description: Added successfully
        "400":
          description: id is invalid (not uuid)
        "401":
          description: Access token is missing or invalid
        "422":
          description: Artist with id doesn't exist
      tags: *a4
    delete:
      operationId: FavoritesController_deleteArtist
      summary: Delete artist from favorites
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        "204":
          description: Deleted successfully
        "400":
          description: id is invalid (not uuid)
        "401":
          description: Access token is missing or invalid
        "404":
          description: Artist was not found
      tags: *a4
...
info:
  title: Home Library Service
  description: Home music library service
  version: 1.0.0
  contact: {}
...
servers: []
components:
  schemas:
...
    Artist:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 2
          example: Freddie Mercury
        grammy:
          type: boolean
          example: false
      required:
        - id
        - name
        - grammy
    CreateArtistDto:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          example: Freddie Mercury
        grammy:
          type: boolean
          example: false
      required:
        - name
        - grammy
    UpdateArtistDto:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          example: Freddie Mercury
        grammy:
          type: boolean
          example: false
    Album:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 2
          example: Innuendo
        year:
          type: number
          minimum: 1900
          maximum: 2023
          example: 2020
        artistId:
          type: string
          nullable: true
          format: uuid
      required:
        - id
        - name
        - year
        - artistId
    CreateAlbumDto:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          example: Innuendo
        year:
          type: number
          minimum: 1900
          maximum: 2023
          example: 2020
        artistId:
          type: string
          nullable: true
          format: uuid
      required:
        - name
        - year
        - artistId
    UpdateAlbumDto:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          example: Innuendo
        year:
          type: number
          minimum: 1900
          maximum: 2023
          example: 2020
        artistId:
          type: string
          nullable: true
          format: uuid
...